Seperate settings for local, production & Travis
The project settings usually reside in the project configuration
folder(with the same name as the project) in a python module known as
settings.py. This is inefficient if different settings are to be used
for different build environments. A good example of this would be the
'DEBUG' setting in the settings.py module. We usually set 'DEBUG=True'
during development and sometimes testing to use the helpful debug
messages of the lightweight development server that comes bundled with
Django, but due to security reasons, it is not advised to leave it at
'True' in a production environment. Or, you may use MySQL as your
database during the development phase but may want to use a lightweight
server like sqlite for testing. Having seperates files for local
development, production and for the Travis CI server lets us do just
that.

The files added are:
- 'base.py': settings.py module is moved into the 'settings' directory
  as 'base.py' and contains global settings which are common in all
  environments
- 'local.py': contains settings specific to the local development and
  testing environments
- 'travis.py': contains settings specific to the Travis CI server for
  its test builds
- 'production.py': contains settings specific to a production
  environment like Heroku

Resolves #25

Signed-off-by: Nitin Paul <npaul2811@gmail.com>