add heroku configuration with production settings and requirements
this adds the configuration for a development server located at
http://contributr.herokuapp.com.

It uses the production settings at contributr.settings.production
set by the DJANGO_SETTINGS_MODULE environment variable.

The requirements and wsgi configuration required in heroku's
environment are added. gunicorn is the http server, psycopg2 is
a database adapter for Postgres, and the other requirements are
used by Heroku to properly communicate with Django.

Procfile and requirements should be there in the root of our
project, so a --pythonpath is required for gunicorn to go in the
right path and find the wsgi file. Also, a production requirements
file is loaded from requirements.txt