add automatic deploying of master branch to heroku
This adds automatic deployment to heroku after tests have passed
in Travis CI for each commit. The API KEY from heroku has been
encrypted for security purposes. "app" parameter is used for
specifying the app that is to be deployed.

Also, previously the .travis.yml had 4 spaced indentation, whereas
it should be 2 spaced. This has been fixed in this commit.