Add blog app with simple view, necessary settings, url-patterns
This is the first skeleton of the blog app. From the url-root
all patterns starting with '/blog/' will be redirected to
the blog. As of now the blog only has one simple view that
sends a text as HttpResponse. The reachability of the page
and the unreachability of the subpages is tested. The folder
contriblog is now part of the INSTALLED_APPS.

Part of: #45