Add Post model for basic blog data
This is a Post model for basic data for a blog. This includes, title, author, body, created date, modified date
and published date. The field author has a foreignkey relationship with the user model. created_date has a field option
of auto_now_add which sets the date automatically when created. modified_date has the field option auto_now which sets
the field to now every time the post is saved. Also registered the Post model to the admin site interface, so posts can
be created, edited and deleted.

Added a test for string representation. This test if the string representation in Posts is equal to the
blog title.