Update blog model and add basic views and templates
Things added:
Custom queryset manager, PostQuerySet. This filters and only displays posts that har publish set to true. This makes it easy to create blog drafts, and to publish only when it's done. See https://docs.djangoproject.com/en/1.8/topics/db/managers/#calling-custom-queryset-methods-from-the-manager
for more information.

Meta class for verbose name and a plural name in the admin. Also set the blog post in reverse chronological order with -created_date, so the latest post is displayed first.

Slugfield, for a user friendly url to each post.

PostAdmin to lists fields with title, author and created date in the admin. Also added a prepopulated field which automatically fills the slug with javascript.

Basic templates and class based generic views for the blog.

Some tests to assert that the custom queryset displays posts that has publish set to true.
Closes #54